refactor: replace error boundary file in components dir
